# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Lennox Hale Publishing Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /automation-rules:
    get:
      operationId: listAutomationRules
      tags: [automations]
      summary: List all automation rules
      parameters:
        - name: mailingListId
          in: query
          required: false
          schema:
            type: integer
        - name: triggerType
          in: query
          required: false
          schema:
            type: string
            enum: [new_subscriber, book_published, series_complete, subscriber_tagged, scheduled]
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of automation rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRule"
    post:
      operationId: createAutomationRule
      tags: [automations]
      summary: Create a new automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRuleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automation-rules/{id}:
    get:
      operationId: getAutomationRule
      tags: [automations]
      summary: Get automation rule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Automation rule details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRuleWithLogs"
        "404":
          description: Not found
    put:
      operationId: updateAutomationRule
      tags: [automations]
      summary: Update an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationRuleBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    delete:
      operationId: deleteAutomationRule
      tags: [automations]
      summary: Delete an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /automation-rules/{id}/toggle:
    post:
      operationId: toggleAutomationRule
      tags: [automations]
      summary: Toggle automation rule active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /automation-rules/{id}/execute:
    post:
      operationId: executeAutomationRule
      tags: [automations]
      summary: Manually execute an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationExecutionResult"
  /automation-logs:
    get:
      operationId: listAutomationLogs
      tags: [automations]
      summary: List automation execution logs
      parameters:
        - name: ruleId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of automation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationLog"

  # ── Public Capture ──
...
components:
  schemas:
...
    AutomationRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        triggerType:
          type: string
          enum: [new_subscriber, book_published, series_complete, subscriber_tagged, scheduled]
        triggerConfig:
          type: object
          nullable: true
        actionType:
          type: string
          enum: [send_email, assign_tag, move_to_list, send_lead_magnet, welcome_sequence]
        actionConfig:
          type: object
          nullable: true
        mailingListId:
          type: integer
          nullable: true
        mailingListName:
          type: string
          nullable: true
        emailTemplateId:
          type: integer
          nullable: true
        emailTemplateName:
          type: string
          nullable: true
        isActive:
          type: boolean
        executionCount:
          type: integer
        lastExecutedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, triggerType, actionType, isActive, executionCount, createdAt, updatedAt]
    AutomationRuleWithLogs:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        triggerType:
          type: string
          enum: [new_subscriber, book_published, series_complete, subscriber_tagged, scheduled]
        triggerConfig:
          type: object
          nullable: true
        actionType:
          type: string
          enum: [send_email, assign_tag, move_to_list, send_lead_magnet, welcome_sequence]
        actionConfig:
          type: object
          nullable: true
        mailingListId:
          type: integer
          nullable: true
        mailingListName:
          type: string
          nullable: true
        emailTemplateId:
          type: integer
          nullable: true
        emailTemplateName:
          type: string
          nullable: true
        isActive:
          type: boolean
        executionCount:
          type: integer
        lastExecutedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        recentLogs:
          type: array
          items:
            $ref: "#/components/schemas/AutomationLog"
      required: [id, name, triggerType, actionType, isActive, executionCount, createdAt, updatedAt, recentLogs]
    CreateAutomationRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        triggerType:
          type: string
          enum: [new_subscriber, book_published, series_complete, subscriber_tagged, scheduled]
        triggerConfig:
          type: object
          nullable: true
        actionType:
          type: string
          enum: [send_email, assign_tag, move_to_list, send_lead_magnet, welcome_sequence]
        actionConfig:
          type: object
          nullable: true
        mailingListId:
          type: integer
          nullable: true
        emailTemplateId:
          type: integer
          nullable: true
        isActive:
          type: boolean
      required: [name, triggerType, actionType]
    UpdateAutomationRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        triggerType:
          type: string
          enum: [new_subscriber, book_published, series_complete, subscriber_tagged, scheduled]
        triggerConfig:
          type: object
          nullable: true
        actionType:
          type: string
          enum: [send_email, assign_tag, move_to_list, send_lead_magnet, welcome_sequence]
        actionConfig:
          type: object
          nullable: true
        mailingListId:
          type: integer
          nullable: true
        emailTemplateId:
          type: integer
          nullable: true
        isActive:
          type: boolean
    AutomationLog:
      type: object
      properties:
        id:
          type: integer
        ruleId:
          type: integer
        ruleName:
          type: string
        subscriberId:
          type: integer
          nullable: true
        subscriberEmail:
          type: string
          nullable: true
        status:
          type: string
          enum: [success, failed, skipped]
        action:
          type: string
        details:
          type: object
          nullable: true
        executedAt:
          type: string
          format: date-time
      required: [id, ruleId, ruleName, status, action, executedAt]
    AutomationExecutionResult:
      type: object
      properties:
        executed:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AutomationLog"
      required: [executed, succeeded, failed, skipped, logs]

    # ── Capture ──
...